chore: refactor devices into separate module
