+++ src/apps/main/device/handlers.ts
@@ -2,25 +2,22 @@ import { ipcMain } from 'electron';
 
 import {
   addBackup,
-  addUnknownDeviceIssue,
   changeBackupPath,
   deleteBackup,
   disableBackup,
   deleteBackupsFromDevice,
   getBackupsFromDevice,
   getPathFromDialog,
-  getOrCreateDevice,
-  getDevices,
-  renameDevice,
   createBackupsFromLocalPaths,
   downloadBackup,
 } from './service';
+import { DeviceModule } from '@/backend/features/backups/device/device.module';
 
-ipcMain.handle('get-or-create-device', getOrCreateDevice);
+ipcMain.handle('get-or-create-device', DeviceModule.getOrCreateDevice);
 
-ipcMain.handle('rename-device', (_, v) => renameDevice(v));
+ipcMain.handle('rename-device', (_, v) => DeviceModule.renameDevice(v));
 
-ipcMain.handle('devices.get-all', () => getDevices());
+ipcMain.handle('devices.get-all', () => DeviceModule.getDevices());
 
 ipcMain.handle('get-backups-from-device', (_, d, c?) => getBackupsFromDevice(d, c));
 
@@ -38,6 +35,6 @@ ipcMain.handle('download-backup', (_, v, fd) => downloadBackup(v, fd));
 
 ipcMain.handle('change-backup-path', (_, v) => changeBackupPath(v));
 
-ipcMain.on('add-device-issue', (_, e) => addUnknownDeviceIssue(e));
+ipcMain.on('add-device-issue', (_, e) => DeviceModule.addUnknownDeviceIssue(e));
 
 ipcMain.handle('get-folder-path', () => getPathFromDialog());
