+++ src/backend/features/backups/device/in/get-device-uuid.ts
@@ -1,0 +1,11 @@
+import configStore from '@/apps/main/config';
+
+export function getDeviceUuid(): string {
+  const deviceUuid = configStore.get('deviceUuid');
+
+  if (deviceUuid === '') {
+    throw new Error('deviceUuid is not defined');
+  }
+
+  return deviceUuid;
+}
