+++ src/backend/features/backups/device/in/save-device-to-config.ts
@@ -1,0 +1,8 @@
+import configStore from '@/apps/main/config';
+import { Device } from '@/apps/main/device/service';
+
+export function saveDeviceToConfig(device: Device) {
+  configStore.set('deviceId', device.id);
+  configStore.set('deviceUuid', device.uuid);
+  configStore.set('backupList', {});
+}
